requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
asttokens==3.0.0
    # via stack-data
audioread==3.0.1
    # via librosa
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via soundfile
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via audio-pond (pyproject.toml)
comm==0.2.2
    # via ipykernel
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
debugpy==1.8.12
    # via ipykernel
decorator==5.1.1
    # via
    #   ipython
    #   librosa
executing==2.2.0
    # via stack-data
filelock==3.17.0
    # via torch
fonttools==4.56.0
    # via matplotlib
fsspec==2025.2.0
    # via torch
idna==3.10
    # via requests
ipykernel==6.29.5
    # via audio-pond (pyproject.toml)
ipython==8.32.0
    # via ipykernel
jedi==0.19.2
    # via ipython
jinja2==3.1.5
    # via torch
joblib==1.4.2
    # via
    #   librosa
    #   scikit-learn
jupyter-client==8.6.3
    # via ipykernel
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
kiwisolver==1.4.8
    # via matplotlib
lazy-loader==0.4
    # via librosa
librosa==0.10.2.post1
    # via
    #   audio-pond (pyproject.toml)
    #   piano-transcription-inference
    #   torchlibrosa
llvmlite==0.44.0
    # via numba
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.0
    # via piano-transcription-inference
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mido==1.3.3
    # via
    #   audio-pond (pyproject.toml)
    #   piano-transcription-inference
mpmath==1.3.0
    # via sympy
msgpack==1.1.0
    # via librosa
nest-asyncio==1.6.0
    # via ipykernel
networkx==3.4.2
    # via torch
numba==0.61.0
    # via librosa
numpy==1.26.4
    # via
    #   contourpy
    #   librosa
    #   matplotlib
    #   numba
    #   scikit-learn
    #   scipy
    #   soundfile
    #   soxr
    #   torchlibrosa
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.2
    # via torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
packaging==24.2
    # via
    #   ipykernel
    #   lazy-loader
    #   matplotlib
    #   mido
    #   pooch
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
piano-transcription-inference==0.0.6
    # via audio-pond (pyproject.toml)
pillow==11.1.0
    # via matplotlib
platformdirs==4.3.6
    # via
    #   jupyter-core
    #   pooch
pooch==1.8.2
    # via librosa
prompt-toolkit==3.0.50
    # via ipython
psutil==6.1.1
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pycparser==2.22
    # via cffi
pydub==0.25.1
    # via audio-pond (pyproject.toml)
pygments==2.19.1
    # via ipython
pyparsing==3.2.1
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   jupyter-client
    #   matplotlib
python-dotenv==1.0.1
    # via audio-pond (pyproject.toml)
pyzmq==26.2.1
    # via
    #   ipykernel
    #   jupyter-client
requests==2.32.3
    # via pooch
scikit-learn==1.6.1
    # via librosa
scipy==1.15.1
    # via
    #   librosa
    #   scikit-learn
six==1.17.0
    # via python-dateutil
soundfile==0.13.1
    # via librosa
soxr==0.5.0.post1
    # via librosa
stack-data==0.6.3
    # via ipython
sympy==1.13.1
    # via torch
threadpoolctl==3.5.0
    # via scikit-learn
torch==2.6.0
    # via audio-pond (pyproject.toml)
torchlibrosa==0.1.0
    # via piano-transcription-inference
tornado==6.4.2
    # via
    #   ipykernel
    #   jupyter-client
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
triton==3.2.0
    # via torch
typing-extensions==4.12.2
    # via
    #   ipython
    #   librosa
    #   torch
urllib3==2.3.0
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
yt-dlp==2025.1.26
    # via audio-pond (pyproject.toml)
